fix(#411): address vybe PR review — NameError in run-cycle + split alerts
(1) NameError on POST /api/canary/run-cycle: handler referenced `valid_ids`
which never existed (commit 02cfc975 only fixed the call site, not the
two trailing usages). Endpoint crashed on every request. Replace with
`requested_ids` (provably valid post-422 check).

(2) services/canary_service.py was 692 lines and growing. Extract the
Slack alert surface (constants + `_emit_transition` + `_build_slack_payload`
+ `_render_forensic` + `_format_row_refs` + `_format_last_red` +
`_render_message`) into services/canary_alerts.py as `CanaryAlerts`.
canary_service.py drops to 347 lines, alerts module is 370 — both well
under the 1000-line target. No behaviour change.

Tests pivoted to `CanaryAlerts._foo` references; 51/51 pass.
